TabooTextInput > suffix, prefix pull up method
