+++ Taboo/src/main/java/com/kwon/taboo/textfield/TabooEditText.kt
@@ -2,19 +2,13 @@ package com.kwon.taboo.textfield
 
 import android.content.Context
 import android.content.res.ColorStateList
-import android.graphics.Typeface
-import android.view.LayoutInflater
 import android.view.View
-import android.view.inputmethod.EditorInfo
-import android.widget.EditText
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.constraintlayout.widget.ConstraintLayout
-import androidx.constraintlayout.widget.ConstraintSet
 import androidx.core.content.ContextCompat
 import com.kwon.taboo.R
 import com.kwon.taboo.enums.AffixType
-import com.kwon.utils.edit.EditorUtils
 
 class TabooEditText(
     private val context: Context,
@@ -45,63 +39,7 @@ class TabooEditText(
      * [affixType]이 [AffixType.SUFFIX]이면 자식 뷰의 개수를 구한 다음에 마지막 자식뷰로 추가한다.
      */
     private fun bindAffixTextView(textView: TextView, affixType: AffixType) {
-        val parentView = view.findViewById<ConstraintLayout>(R.id.cl_edit_text_wrapper)
-        val addIndex = if (affixType == AffixType.PREFIX) 0 else parentView.childCount
-
-        // 자식 뷰에 추가
-        parentView.addView(textView, addIndex)
-
-        ConstraintSet().apply {
-            clone(parentView)
-            applyAffixConstraints(this, textView, affixType, parentView)
-            applyTo(parentView)
-        }
-    }
-
-    /**
-     * [AffixType.PREFIX] 또는 [AffixType.SUFFIX]에 따라 [TextView]의 [ConstraintLayout] 제약 조건을 설정하는 메소드.
-     *
-     * @param constraintSet [ConstraintSet] 인스턴스로, 부모 레이아웃의 제약 조건을 변경하는 데 사용됨.
-     * @param view 제약 조건을 적용할 대상 [TextView].
-     * @param affixType 접두사(PREFIX) 또는 접미사(SUFFIX) 유형을 지정하는 Enum 값. ([AffixType.PREFIX], [AffixType.SUFFIX])
-     * @param parentView 제약 조건을 적용할 부모 [ConstraintLayout].
-     *
-     * - [AffixType.PREFIX] 경우:
-     *   1. `textView`를 부모 뷰의 START와 연결
-     *   2. `textView`의 END를 `editText`의 START와 연결
-     *   3. `editText`의 START를 `textView`의 END와 연결
-     *
-     * - [AffixType.SUFFIX] 경우:
-     *   1. `textView`의 START를 `editText`의 END와 연결
-     *   2. `textView`의 END를 부모 뷰의 END와 연결
-     *   3. `editText`의 END를 `textView`의 START와 연결
-     *
-     * - 공통 설정:
-     *   1. `textView`의 TOP을 부모 뷰의 TOP과 연결
-     *   2. `textView`의 BOTTOM을 부모 뷰의 BOTTOM과 연결
-     */
-    private fun applyAffixConstraints(
-        constraintSet: ConstraintSet,
-        view: View,
-        affixType: AffixType,
-        parentView: ConstraintLayout
-    ) {
-        when (affixType) {
-            AffixType.PREFIX -> {
-                constraintSet.connect(view.id, ConstraintSet.START, parentView.id, ConstraintSet.START)
-                constraintSet.connect(view.id, ConstraintSet.END, editText.id, ConstraintSet.START)
-                constraintSet.connect(editText.id, ConstraintSet.START, view.id, ConstraintSet.END)
-            }
-            AffixType.SUFFIX -> {
-                constraintSet.connect(view.id, ConstraintSet.START, editText.id, ConstraintSet.END)
-                constraintSet.connect(view.id, ConstraintSet.END, parentView.id, ConstraintSet.END)
-                constraintSet.connect(editText.id, ConstraintSet.END, view.id, ConstraintSet.START)
-            }
-        }
-
-        // 공통 제약 조건 추가
-        constraintSet.connect(view.id, ConstraintSet.TOP, parentView.id, ConstraintSet.TOP)
-        constraintSet.connect(view.id, ConstraintSet.BOTTOM, parentView.id, ConstraintSet.BOTTOM)
+        bindView(textView, affixType)
     }
 
     /**
@@ -150,13 +88,14 @@ class TabooEditText(
     private fun updatePasswordToggleEnabled() {
         if (passwordToggleEnable) {
             createPasswordToggleButton()
+
+            bindView(passwordToggleButton!!, AffixType.SUFFIX)
         } else {
             removePasswordToggleButton()
         }
     }
 
     private fun createPasswordToggleButton() {
-        val parentView = view.findViewById<ConstraintLayout>(R.id.cl_edit_text_wrapper)
         passwordToggleButton = ImageView(context).apply {
             setImageDrawable(ContextCompat.getDrawable(context, passwordToggleIcon))
             setColorFilter(ContextCompat.getColor(context, passwordToggleIconActiveColor))
@@ -174,15 +113,6 @@ class TabooEditText(
                 // 비밀번호 표시 상태값 변경
                 isVisiblePassword = !isVisiblePassword
             }
-        }
-
-        // 자식 뷰에 추가
-        parentView.addView(passwordToggleButton, 1)
-
-        ConstraintSet().apply {
-            clone(parentView)
-            applyAffixConstraints(this, passwordToggleButton!!, AffixType.SUFFIX, parentView)
-            applyTo(parentView)
         }
     }
