+++ Taboo/src/main/res/layout/taboo_edit_text.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<merge
+<androidx.constraintlayout.widget.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:background="@drawable/selector_taboo_edit_text"
-    tools:parentTag="androidx.constraintlayout.widget.ConstraintLayout"
+    android:paddingHorizontal="15dp"
     app:layout_constraintTop_toBottomOf="@id/cl_edit_text_title_wrapper"
     app:layout_constraintStart_toStartOf="parent"
     app:layout_constraintEnd_toEndOf="parent">
@@ -25,4 +25,4 @@
         app:layout_constraintTop_toTopOf="parent"
         tools:hint="아이디를 입력해주세요."/>
 
-</merge>
+</androidx.constraintlayout.widget.ConstraintLayout>
