+++ Taboo/src/main/res/layout/taboo_text_input.xml
@@ -40,8 +40,6 @@
             android:id="@+id/cl_edit_text_wrapper"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:background="@drawable/selector_taboo_edit_text"
-            android:paddingHorizontal="15dp"
             android:layout_marginTop="4dp"
             app:layout_constraintTop_toBottomOf="@id/cl_edit_text_title_wrapper"
             app:layout_constraintStart_toStartOf="parent"/>
